Finish casing fix from Wireguard to WireGuard
This also removes from the project various duplicate file entries that
differed only in casing. Note that, if working with a case-insensitive
filesystem (as are generally used on macOS systems), it is recommended
to destroy the affected files in the working tree and check them out
again. Failure to do this will result in XCode recreating duplicate file
entries in the project.